Move to file-scoped namespaces
