+++ src/Spydersoft.Platform.Hosting/Spydersoft.Platform.Hosting.ApiTests/Controllers/CacheTestController.cs
@@ -3,52 +3,51 @@ using Microsoft.AspNetCore.Mvc;
 using Spydersoft.Platform.Hosting.ApiTests.Models;
 using ZiggyCreatures.Caching.Fusion;
 
-namespace Spydersoft.Platform.Hosting.ApiTests.Controllers
+namespace Spydersoft.Platform.Hosting.ApiTests.Controllers;
+
+[ApiController]
+[Route("[controller]")]
+public class CacheTestController(IFusionCache? cache = null) : ControllerBase
 {
-    [ApiController]
-    [Route("[controller]")]
-    public class CacheTestController(IFusionCache? cache = null) : ControllerBase
+    [HttpGet("{id:int}")]
+    public async Task<CacheObjectOne> Get(int id)
     {
-        [HttpGet("{id:int}")]
-        public async Task<CacheObjectOne> Get(int id)
+        if (cache == null)
         {
-            if (cache == null)
+            await Task.Delay(5000);
+            return new CacheObjectOne
             {
-                await Task.Delay(5000);
+                Name = $"Name{id}",
+                Age = id,
+                CreatedAt = DateTime.UtcNow
+            };
+        }
+
+        return await cache.GetOrSetAsync<CacheObjectOne>($"CacheObjectOne:{id}",
+            async _ => {
+                await Task.Delay(2000); // Simulate a delay for cache population
                 return new CacheObjectOne
                 {
                     Name = $"Name{id}",
                     Age = id,
                     CreatedAt = DateTime.UtcNow
                 };
-            }
+            },
+            options => {
+                options.SetDuration(TimeSpan.FromSeconds(10))
+                       .SetFactoryTimeouts(TimeSpan.FromSeconds(5), TimeSpan.FromSeconds(10));
+            });
+    }
 
-            return await cache.GetOrSetAsync<CacheObjectOne>($"CacheObjectOne:{id}",
-                async _ => {
-                    await Task.Delay(2000); // Simulate a delay for cache population
-                    return new CacheObjectOne
-                    {
-                        Name = $"Name{id}",
-                        Age = id,
-                        CreatedAt = DateTime.UtcNow
-                    };
-                },
-                options => {
-                    options.SetDuration(TimeSpan.FromSeconds(10))
-                           .SetFactoryTimeouts(TimeSpan.FromSeconds(5), TimeSpan.FromSeconds(10));
-                });
-        }
+    [HttpGet("HasDistributedCache")]
+    public IActionResult HasDistributedCache()
+    {
+        return Ok(new { HasL2Cache = cache?.HasDistributedCache ?? false });
+    }
 
-        [HttpGet("HasDistributedCache")]
-        public IActionResult HasDistributedCache()
-        {
-            return Ok(new { HasL2Cache = cache?.HasDistributedCache ?? false });
-        }
-
-        [HttpGet("HasCache")]
-        public IActionResult HasCache()
-        {
-            return Ok(new { HasCache = cache != null });
-        }
+    [HttpGet("HasCache")]
+    public IActionResult HasCache()
+    {
+        return Ok(new { HasCache = cache != null });
     }
 }
