+++ src/Spydersoft.Platform.Hosting/Spydersoft.Platform.Hosting.ApiTests/Models/CacheObjectOne.cs
@@ -1,9 +1,8 @@
-﻿namespace Spydersoft.Platform.Hosting.ApiTests.Models
+﻿namespace Spydersoft.Platform.Hosting.ApiTests.Models;
+
+public class CacheObjectOne
 {
-    public class CacheObjectOne
-    {
-        public string? Name { get; set; }
-        public int Age { get; set; }
-        public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
-    }
+    public string? Name { get; set; }
+    public int Age { get; set; }
+    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
 }
