+++ src/Spydersoft.Platform.Hosting/Spydersoft.Platform.Hosting.ApiTests/OptionsTests/NestedOptionSection.cs
@@ -1,12 +1,11 @@
 ﻿using Spydersoft.Platform.Attributes;
 
-namespace Spydersoft.Platform.Hosting.ApiTests.OptionsTests
-{
-    [SpydersoftOptions(nameof(NestedOptionSection), "nested")]
-    public class NestedOptionSection
-    {
-        public string NestedOption1 { get; set; } = "NestedOption1";
+namespace Spydersoft.Platform.Hosting.ApiTests.OptionsTests;
 
-        public string NestedOption2 { get; set; } = "NestedOption2";
-    }
+[SpydersoftOptions(nameof(NestedOptionSection), "nested")]
+public class NestedOptionSection
+{
+    public string NestedOption1 { get; set; } = "NestedOption1";
+
+    public string NestedOption2 { get; set; } = "NestedOption2";
 }
