+++ src/Spydersoft.Platform.Hosting/Spydersoft.Platform.Hosting.ApiTests/OptionsTests/RootOptionSection.cs
@@ -1,12 +1,11 @@
 ﻿using Spydersoft.Platform.Attributes;
 
-namespace Spydersoft.Platform.Hosting.ApiTests.OptionsTests
-{
-    [SpydersoftOptions(nameof(RootOptionSection), "root")]
-    public class RootOptionSection
-    {
-        public string Option1 { get; set; } = "Option1";
+namespace Spydersoft.Platform.Hosting.ApiTests.OptionsTests;
 
-        public string Option2 { get; set; } = "Option2";
-    }
+[SpydersoftOptions(nameof(RootOptionSection), "root")]
+public class RootOptionSection
+{
+    public string Option1 { get; set; } = "Option1";
+
+    public string Option2 { get; set; } = "Option2";
 }
