+++ src/Spydersoft.Platform.Hosting/Spydersoft.Platform.Hosting.ApiTests/OptionsTests/TaggedNotLoadedOptions.cs
@@ -1,10 +1,9 @@
 ﻿using Spydersoft.Platform.Attributes;
 
-namespace Spydersoft.Platform.Hosting.ApiTests.OptionsTests
+namespace Spydersoft.Platform.Hosting.ApiTests.OptionsTests;
+
+//[SpydersoftOptions(nameof(TaggedNotLoadedOptions), "notloaded")]
+public class TaggedNotLoadedOptions
 {
-    //[SpydersoftOptions(nameof(TaggedNotLoadedOptions), "notloaded")]
-    public class TaggedNotLoadedOptions
-    {
-        public string NotLoadedOption1 { get; set; } = "NotLoadedOption1";
-    }
+    public string NotLoadedOption1 { get; set; } = "NotLoadedOption1";
 }
