+++ src/Spydersoft.Platform.Hosting/Spydersoft.Platform.Hosting.UnitTests/ApiTests/Options/ConfiguredOptionsTests.cs
@@ -8,72 +8,71 @@ using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
 
-namespace Spydersoft.Platform.Hosting.UnitTests.ApiTests.Options
+namespace Spydersoft.Platform.Hosting.UnitTests.ApiTests.Options;
+
+public class ConfiguredOptionsTests : ApiTestBase
 {
-    public class ConfiguredOptionsTests : ApiTestBase
+    public override string Environment => "Options";
+
+    [Test]
+    public async Task RootOptions()
     {
-        public override string Environment => "Options";
+        var result = await Client.GetAsync($"options/root");
 
-        [Test]
-        public async Task RootOptions()
+        using var jsonResult = JsonDocument.Parse(await result.Content.ReadAsStringAsync());
+
+        var telemetryNode = jsonResult.RootElement;
+
+        var details = telemetryNode.Deserialize<RootOptionSection>(
+                JsonOptions
+        );
+        Assert.Multiple(() =>
         {
-            var result = await Client.GetAsync($"options/root");
+            Assert.That(result.StatusCode, Is.EqualTo(HttpStatusCode.OK));
+            Assert.That(details, Is.Not.Null);
+            Assert.That(details?.Option1, Is.EqualTo("Value1FromOptions"));
+            Assert.That(details?.Option2, Is.EqualTo("Value2FromOptions"));
+        });
+    }
 
-            using var jsonResult = JsonDocument.Parse(await result.Content.ReadAsStringAsync());
+    [Test]
+    public async Task NestedOptions()
+    {
+        var result = await Client.GetAsync($"options/nested");
 
-            var telemetryNode = jsonResult.RootElement;
+        using var jsonResult = JsonDocument.Parse(await result.Content.ReadAsStringAsync());
 
-            var details = telemetryNode.Deserialize<RootOptionSection>(
-                    JsonOptions
-            );
-            Assert.Multiple(() =>
-            {
-                Assert.That(result.StatusCode, Is.EqualTo(HttpStatusCode.OK));
-                Assert.That(details, Is.Not.Null);
-                Assert.That(details?.Option1, Is.EqualTo("Value1FromOptions"));
-                Assert.That(details?.Option2, Is.EqualTo("Value2FromOptions"));
-            });
-        }
+        var telemetryNode = jsonResult.RootElement;
 
-        [Test]
-        public async Task NestedOptions()
+        var details = telemetryNode.Deserialize<NestedOptionSection>(
+                JsonOptions
+        );
+        Assert.Multiple(() =>
         {
-            var result = await Client.GetAsync($"options/nested");
+            Assert.That(result.StatusCode, Is.EqualTo(HttpStatusCode.OK));
+            Assert.That(details, Is.Not.Null);
+            Assert.That(details?.NestedOption1, Is.EqualTo("NestedValue1FromOptions"));
+            Assert.That(details?.NestedOption2, Is.EqualTo("NestedValue2FromOptions"));
+        });
+    }
 
-            using var jsonResult = JsonDocument.Parse(await result.Content.ReadAsStringAsync());
+    [Test]
+    public async Task NotLoadedOptions()
+    {
+        var result = await Client.GetAsync($"options/notloaded");
 
-            var telemetryNode = jsonResult.RootElement;
+        using var jsonResult = JsonDocument.Parse(await result.Content.ReadAsStringAsync());
 
-            var details = telemetryNode.Deserialize<NestedOptionSection>(
-                    JsonOptions
-            );
-            Assert.Multiple(() =>
-            {
-                Assert.That(result.StatusCode, Is.EqualTo(HttpStatusCode.OK));
-                Assert.That(details, Is.Not.Null);
-                Assert.That(details?.NestedOption1, Is.EqualTo("NestedValue1FromOptions"));
-                Assert.That(details?.NestedOption2, Is.EqualTo("NestedValue2FromOptions"));
-            });
-        }
+        var telemetryNode = jsonResult.RootElement;
 
-        [Test]
-        public async Task NotLoadedOptions()
+        var details = telemetryNode.Deserialize<TaggedNotLoadedOptions>(
+                JsonOptions
+        );
+        Assert.Multiple(() =>
         {
-            var result = await Client.GetAsync($"options/notloaded");
-
-            using var jsonResult = JsonDocument.Parse(await result.Content.ReadAsStringAsync());
-
-            var telemetryNode = jsonResult.RootElement;
-
-            var details = telemetryNode.Deserialize<TaggedNotLoadedOptions>(
-                    JsonOptions
-            );
-            Assert.Multiple(() =>
-            {
-                Assert.That(result.StatusCode, Is.EqualTo(HttpStatusCode.OK));
-                Assert.That(details, Is.Not.Null);
-                Assert.That(details?.NotLoadedOption1, Is.EqualTo("NotLoadedOption1"));
-            });
-        }
+            Assert.That(result.StatusCode, Is.EqualTo(HttpStatusCode.OK));
+            Assert.That(details, Is.Not.Null);
+            Assert.That(details?.NotLoadedOption1, Is.EqualTo("NotLoadedOption1"));
+        });
     }
 }
