+++ src/Spydersoft.Platform.Hosting/Spydersoft.Platform.Hosting/StartupExtensions/FusionCacheExtensions.cs
@@ -11,112 +11,111 @@ using ZiggyCreatures.Caching.Fusion;
 using ZiggyCreatures.Caching.Fusion.Backplane.StackExchangeRedis;
 using ZiggyCreatures.Caching.Fusion.Serialization.SystemTextJson;
 
-namespace Spydersoft.Platform.Hosting.StartupExtensions
+namespace Spydersoft.Platform.Hosting.StartupExtensions;
+
+public static class FusionCacheExtensions
 {
-    public static class FusionCacheExtensions
+    /// <summary>
+    /// Adds the aspire fusion cache.
+    /// </summary>
+    /// <param name="builder">The builder.</param>
+    /// <returns>WebApplicationBuilder.</returns>
+    public static WebApplicationBuilder AddSpydersoftFusionCache(this WebApplicationBuilder builder, Action<IFusionCacheBuilder>? additionalConfiguration = null)
     {
-        /// <summary>
-        /// Adds the aspire fusion cache.
-        /// </summary>
-        /// <param name="builder">The builder.</param>
-        /// <returns>WebApplicationBuilder.</returns>
-        public static WebApplicationBuilder AddSpydersoftFusionCache(this WebApplicationBuilder builder, Action<IFusionCacheBuilder>? additionalConfiguration = null)
+        var cacheOptions = new FusionCacheConfigOptions();
+        builder.Configuration.GetSection(FusionCacheConfigOptions.SectionName).Bind(cacheOptions);
+
+        builder.Services.ConfigureFusionCache(cacheOptions, additionalConfiguration);
+
+        return builder;
+    }
+
+    /// <summary>
+    /// Configures the fusion cache.
+    /// </summary>
+    /// <param name="services">The services.</param>
+    /// <param name="options">The options.</param>
+    /// <returns>IServiceCollection.</returns>
+    public static IServiceCollection ConfigureFusionCache(this IServiceCollection services, FusionCacheConfigOptions options, Action<IFusionCacheBuilder>? additionalConfiguration = null)
+    {
+        if (!options.Enabled)
         {
-            var cacheOptions = new FusionCacheConfigOptions();
-            builder.Configuration.GetSection(FusionCacheConfigOptions.SectionName).Bind(cacheOptions);
-
-            builder.Services.ConfigureFusionCache(cacheOptions, additionalConfiguration);
-
-            return builder;
-        }
-
-        /// <summary>
-        /// Configures the fusion cache.
-        /// </summary>
-        /// <param name="services">The services.</param>
-        /// <param name="options">The options.</param>
-        /// <returns>IServiceCollection.</returns>
-        public static IServiceCollection ConfigureFusionCache(this IServiceCollection services, FusionCacheConfigOptions options, Action<IFusionCacheBuilder>? additionalConfiguration = null)
-        {
-            if (!options.Enabled)
-            {
-                return services;
-            }
-
-            services.AddMemoryCache();
-
-            // Could provide options.CacheName for keyed services, but this only 
-            //  supports a single FusionCache instance per application now, not 
-            //  necessary to make it keyed.
-            var fusionCache = services.AddFusionCache()
-                .WithDefaultEntryOptions(new FusionCacheEntryOptions
-                {
-                    Duration = TimeSpan.FromMinutes(options.DefaultEntryDurationInMinutes),
-
-                    IsFailSafeEnabled = options.EnableFailSafe,
-                    FailSafeMaxDuration = TimeSpan.FromMinutes(options.FailSafeMaxMinutes),
-                    FailSafeThrottleDuration = TimeSpan.FromMinutes(options.FailSafeThrottleMinutes),
-
-                    FactorySoftTimeout = TimeSpan.FromMilliseconds(options.FactoryTimeoutSoftMs),
-                    FactoryHardTimeout = TimeSpan.FromMilliseconds(options.FactoryTimeoutHardMs),
-
-                    DistributedCacheSoftTimeout = TimeSpan.FromMinutes(options.DistributedCacheSoftTimeoutMinutes),
-                    DistributedCacheHardTimeout = TimeSpan.FromMinutes(options.DistributedCacheHardTimeoutMinutes),
-                    AllowBackgroundDistributedCacheOperations = true,
-                });
-
-
-            switch (options.DistributedCacheType)
-            {
-                case CacheType.Memory:
-                    services.AddDistributedMemoryCache();
-                    break;
-                case CacheType.Redis:
-                    if (string.IsNullOrWhiteSpace(options.Redis.ConnectionString))
-                    {
-                        throw new ConfigurationException("Redis connection string must be provided when using Redis cache.");
-                    }
-                    fusionCache.ConfigureRedisWithBackplane(options);
-
-                    break;
-                case CacheType.None:
-                    // No distributed cache configured
-                    break;
-                default:
-                    throw new NotSupportedException($"Distributed cache type '{options.DistributedCacheType}' is not supported.");
-            }
-
-            if (additionalConfiguration != null)
-            {
-                additionalConfiguration(fusionCache);
-            }
-
             return services;
         }
 
-        [ExcludeFromCodeCoverage(Justification = "Requires Integtation Tests to verify Redis configuration")]
-        private static void ConfigureRedisWithBackplane(this IFusionCacheBuilder fusionCache, FusionCacheConfigOptions options)
+        services.AddMemoryCache();
+
+        // Could provide options.CacheName for keyed services, but this only 
+        //  supports a single FusionCache instance per application now, not 
+        //  necessary to make it keyed.
+        var fusionCache = services.AddFusionCache()
+            .WithDefaultEntryOptions(new FusionCacheEntryOptions
+            {
+                Duration = TimeSpan.FromMinutes(options.DefaultEntryDurationInMinutes),
+
+                IsFailSafeEnabled = options.EnableFailSafe,
+                FailSafeMaxDuration = TimeSpan.FromMinutes(options.FailSafeMaxMinutes),
+                FailSafeThrottleDuration = TimeSpan.FromMinutes(options.FailSafeThrottleMinutes),
+
+                FactorySoftTimeout = TimeSpan.FromMilliseconds(options.FactoryTimeoutSoftMs),
+                FactoryHardTimeout = TimeSpan.FromMilliseconds(options.FactoryTimeoutHardMs),
+
+                DistributedCacheSoftTimeout = TimeSpan.FromMinutes(options.DistributedCacheSoftTimeoutMinutes),
+                DistributedCacheHardTimeout = TimeSpan.FromMinutes(options.DistributedCacheHardTimeoutMinutes),
+                AllowBackgroundDistributedCacheOperations = true,
+            });
+
+
+        switch (options.DistributedCacheType)
         {
-            // ADD JSON.NET BASED SERIALIZATION FOR FUSION CACHE
-            fusionCache
-            .WithSerializer(
-                new FusionCacheSystemTextJsonSerializer(new JsonSerializerOptions
+            case CacheType.Memory:
+                services.AddDistributedMemoryCache();
+                break;
+            case CacheType.Redis:
+                if (string.IsNullOrWhiteSpace(options.Redis.ConnectionString))
                 {
-                    PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-                    WriteIndented = false,
-                    DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
-                }))
-            .WithDistributedCache(new RedisCache(
-                            new RedisCacheOptions()
-                            {
-                                Configuration = options.Redis.ConnectionString,
-                                InstanceName = options.Redis.InstanceName
-                            }
-            ))
-            // ADD THE FUSION CACHE BACKPLANE FOR REDIS
-            .WithBackplane(
-                new RedisBackplane(new RedisBackplaneOptions() { Configuration = options.Redis.ConnectionString })
-            );
+                    throw new ConfigurationException("Redis connection string must be provided when using Redis cache.");
+                }
+                fusionCache.ConfigureRedisWithBackplane(options);
+
+                break;
+            case CacheType.None:
+                // No distributed cache configured
+                break;
+            default:
+                throw new NotSupportedException($"Distributed cache type '{options.DistributedCacheType}' is not supported.");
         }
+
+        if (additionalConfiguration != null)
+        {
+            additionalConfiguration(fusionCache);
+        }
+
+        return services;
+    }
+
+    [ExcludeFromCodeCoverage(Justification = "Requires Integtation Tests to verify Redis configuration")]
+    private static void ConfigureRedisWithBackplane(this IFusionCacheBuilder fusionCache, FusionCacheConfigOptions options)
+    {
+        // ADD JSON.NET BASED SERIALIZATION FOR FUSION CACHE
+        fusionCache
+        .WithSerializer(
+            new FusionCacheSystemTextJsonSerializer(new JsonSerializerOptions
+            {
+                PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+                WriteIndented = false,
+                DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
+            }))
+        .WithDistributedCache(new RedisCache(
+                        new RedisCacheOptions()
+                        {
+                            Configuration = options.Redis.ConnectionString,
+                            InstanceName = options.Redis.InstanceName
+                        }
+        ))
+        // ADD THE FUSION CACHE BACKPLANE FOR REDIS
+        .WithBackplane(
+            new RedisBackplane(new RedisBackplaneOptions() { Configuration = options.Redis.ConnectionString })
+        );
     }
 }
