+++ src/Spydersoft.Platform/Spydersoft.Platform.UnitTests/ExceptionsTests/ConfigurationExceptionTests.cs
@@ -1,15 +1,14 @@
 using Spydersoft.Platform.Exceptions;
 
-namespace Spydersoft.Platform.UnitTests.ExceptionTests
+namespace Spydersoft.Platform.UnitTests.ExceptionTests;
+
+public class Tests
 {
-    public class Tests
+    [Test]
+    public void Validate_Constructor()
     {
-        [Test]
-        public void Validate_Constructor()
-        {
-            var message = "My Configuration Message;";
-            var exception = new ConfigurationException(message);
-            Assert.That(exception.Message, Is.EqualTo(message));
-        }
+        var message = "My Configuration Message;";
+        var exception = new ConfigurationException(message);
+        Assert.That(exception.Message, Is.EqualTo(message));
     }
 }
