+++ src/Spydersoft.Platform/Spydersoft.Platform/Attributes/SpydersoftOptionsAttribute.cs
@@ -1,20 +1,19 @@
 ﻿using System;
 
-namespace Spydersoft.Platform.Attributes
+namespace Spydersoft.Platform.Attributes;
+
+[AttributeUsage(AttributeTargets.Class)]
+public class SpydersoftOptionsAttribute : Attribute
 {
-    [AttributeUsage(AttributeTargets.Class)]
-    public class SpydersoftOptionsAttribute : Attribute
+    public SpydersoftOptionsAttribute(string sectionName, string tags = "")
     {
-        public SpydersoftOptionsAttribute(string sectionName, string tags = "")
-        {
-            SectionName = sectionName;
-            RawTags = tags;
-            Tags = tags.Split(',', StringSplitOptions.RemoveEmptyEntries);
-        }
-        public string SectionName { get; }
-
-        public string RawTags { get; }
-
-        public string[] Tags { get; }
+        SectionName = sectionName;
+        RawTags = tags;
+        Tags = tags.Split(',', StringSplitOptions.RemoveEmptyEntries);
     }
+    public string SectionName { get; }
+
+    public string RawTags { get; }
+
+    public string[] Tags { get; }
 }
