+++ src/Spydersoft.Platform/Spydersoft.Platform/Exceptions/ConfigurationException.cs
@@ -1,12 +1,10 @@
 ﻿using System;
 
-namespace Spydersoft.Platform.Exceptions
-{
-    public class ConfigurationException : Exception
-    {
-        public ConfigurationException(string message) : base(message)
-        {
-        }
-    }
+namespace Spydersoft.Platform.Exceptions;
 
+public class ConfigurationException : Exception
+{
+    public ConfigurationException(string message) : base(message)
+    {
+    }
 }
